Fixed: * PIMS-1100 Expiring Stock Report > Allow user to filter by category * PIMS-1108 Certificate of Donation > Date shown should be expected shipping date, not expected delivery date * PIMS-1107 Certificate of Donation > Contains duplicate date value on second page * PIMS-1092 Order summary on the dashboard shows incorrect totals * PIMS-1097 Inventory Browser > Rows should be same height as rows in the Expiring Stock report * PIMS-1094 Add to Shipment > Need to include expiration date in order to allow Blake/admin to choose items based on FEFO policy * PIMS-1103 Daily Transaction Report > Transaction ID should link to transaction/shipment associated with the transaction PIMS-1096 Inventory Browser > Add "Pending" to Qty In / Qty Out * Upgrade to the latest version of the docx4j library * Refactored some of the Document code (doc4j controller, document service)
Working on: 
* PIMS-1112 Allow user to fulfill a request


git-svn-id: https://dev.pih-emr.org/svn/repository/projects/warehouse/trunk@10554 f0ac6d2b-dfc5-4988-9158-fc266d298d87

+++ grails-app/controllers/org/pih/warehouse/shipping/CreateShipmentWorkflowController.groovy
@@ -23,7 +23,7 @@ class CreateShipmentWorkflowController {
     			
     			// create a new shipment instance if we don't have one already
     			if (!flow.shipmentInstance) { 
-    				flow.shipmentInstance = shipmentService.getShipmentInstance(params.id)
+    				flow.shipmentInstance = shipmentService.getShipmentInstance(params.id as Long)
     				flow.shipmentWorkflow = shipmentService.getShipmentWorkflow(flow.shipmentInstance)
     			}
 				if (params.skipTo) {
